+++ fluentui_topappbars/src/main/java/com/microsoft/fluentui/tokenized/AppBar.kt
@@ -47,132 +47,8 @@ import androidx.compose.ui.unit.*
 import androidx.compose.ui.window.Popup
 import androidx.compose.ui.window.PopupPositionProvider
 import androidx.compose.ui.window.PopupProperties
+import com.microsoft.fluentui.util.clickableWithTooltip
 import kotlinx.coroutines.delay
-
-@OptIn(ExperimentalFoundationApi::class)
-private fun Modifier.clickableWithTooltip(
-    tooltipText: String,
-    tooltipEnabled: Boolean = false,
-    backgroundColor: Brush = SolidColor(Color.Unspecified),
-    cornerRadius: Dp = 8.dp,
-    textStyle: TextStyle = TextStyle(color = Color.Black, fontWeight = FontWeight.Normal),
-    padding: Dp = 12.dp,
-    offset: DpOffset = DpOffset(0.dp, 0.dp),
-    timeout: Long = 2000L,
-    showRippleOnClick: Boolean = true,
-    clickRippleColor: Color = Color.Unspecified,
-    onClick: (() -> Unit)? = null,
-    onLongClick: (() -> Unit)? = null
-): Modifier = composed {
-    var isTooltipVisible by remember { mutableStateOf(false) }
-    var anchorBounds by remember { mutableStateOf(IntRect.Zero) }
-    val interactionSource = remember { MutableInteractionSource() }
-
-    val currentOnClick by rememberUpdatedState(onClick)
-    val currentOnLongClick by rememberUpdatedState(onLongClick)
-    val currentTooltipEnabled by rememberUpdatedState(tooltipEnabled)
-
-    val clickableModifier = Modifier.combinedClickable(
-        interactionSource = interactionSource,
-        indication = if (showRippleOnClick) rememberRipple(color = clickRippleColor) else null,
-        onClick = { currentOnClick?.invoke() },
-        onLongClick = {
-            currentOnLongClick?.invoke()
-            if (currentTooltipEnabled) {
-                isTooltipVisible = true
-            }
-        }
-    )
-
-    val positionModifier = Modifier.onGloballyPositioned { layoutCoordinates ->
-        anchorBounds = IntRect(
-            left = layoutCoordinates.positionInWindow().x.toInt(),
-            top = layoutCoordinates.positionInWindow().y.toInt(),
-            right = (layoutCoordinates.positionInWindow().x + layoutCoordinates.size.width).toInt(),
-            bottom = (layoutCoordinates.positionInWindow().y + layoutCoordinates.size.height).toInt()
-        )
-    }
-
-    if (isTooltipVisible) {
-        Tooltip(
-            tooltipText = tooltipText,
-            backgroundColor = backgroundColor,
-            cornerRadius = cornerRadius,
-            textStyle = textStyle,
-            padding = padding,
-            offset = offset,
-            onDismissRequest = { isTooltipVisible = false }
-        )
-
-        LaunchedEffect(isTooltipVisible, timeout) {
-            delay(timeout)
-            isTooltipVisible = false
-        }
-    }
-
-    this
-        .then(if (tooltipEnabled) positionModifier else Modifier)
-        .then(clickableModifier)
-}
-
-@Composable
-private fun Tooltip(
-    tooltipText: String,
-    backgroundColor: Brush,
-    cornerRadius: Dp,
-    textStyle: TextStyle,
-    padding: Dp,
-    offset: DpOffset,
-    onDismissRequest: () -> Unit
-) {
-    val density = LocalDensity.current
-    val positionProvider = remember(offset) {
-        TooltipPositionProvider(
-            offset = offset,
-            density = density
-        )
-    }
-    Popup(
-        popupPositionProvider = positionProvider,
-        onDismissRequest = onDismissRequest,
-        properties = PopupProperties(
-            dismissOnBackPress = true,
-            dismissOnClickOutside = true,
-            focusable = false,
-        )
-    ) {
-        Column(
-            modifier = Modifier.shadow(4.dp, clip = false)
-        ) {
-            Box(
-                modifier = Modifier
-                    .background(backgroundColor, RoundedCornerShape(cornerRadius))
-                    .padding(padding)
-            ) {
-                Text(text = tooltipText, style = textStyle)
-            }
-        }
-    }
-}
-
-private class TooltipPositionProvider(
-    private val offset: DpOffset,
-    private val density: Density
-) : PopupPositionProvider {
-    override fun calculatePosition(
-        anchorBounds: IntRect,
-        windowSize: IntSize,
-        layoutDirection: LayoutDirection,
-        popupContentSize: IntSize
-    ): IntOffset {
-        val offsetX = with(density) { offset.x.roundToPx() }
-        val offsetY = with(density) { offset.y.roundToPx() }
-        val y =
-            if ((anchorBounds.top - popupContentSize.height) > 0) anchorBounds.top - popupContentSize.height + offsetY else anchorBounds.bottom + offsetY
-        val x = (anchorBounds.left + (anchorBounds.width - popupContentSize.width) / 2) + offsetX
-        return IntOffset(x, y)
-    }
-}
 
 /**
  * An app bar appears at the top of an app screen, below the status bar,
